requirements.txt
Flask
gunicorn
langchain
openai
python-dotenv
pipedrive-python-lib